% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_write_outputs.R
\name{.add_date_axis}
\alias{.add_date_axis}
\title{Add a date axis with ~5 evenly spaced labels}
\usage{
.add_date_axis(x)
}
\arguments{
\item{x}{IDate or Date vector used as x-axis values}
}
\description{
Suppresses default x-axis and adds date labels (including first and last date).
}
\keyword{internal}
