% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.add_stage_fcr_qc}
\alias{.add_stage_fcr_qc}
\title{Add weight stage FCR quality control}
\usage{
.add_stage_fcr_qc(daily_data, phenotypes, cfg)
}
\description{
Add weight stage FCR quality control
}
\keyword{internal}
