% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_base_phenotypes}
\alias{.calc_base_phenotypes}
\title{Calculate base phenotypes (standard Report format)}
\usage{
.calc_base_phenotypes(daily_records)
}
\description{
Calculate base phenotypes (standard Report format)
}
\keyword{internal}
