% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_phenotypes_by_age_stages}
\alias{.calc_phenotypes_by_age_stages}
\title{Calculate phenotypes by age stages}
\usage{
.calc_phenotypes_by_age_stages(
  daily_records,
  phenotype_method,
  cfg,
  target_age_stages
)
}
\description{
Calculate phenotypes by age stages
}
\keyword{internal}
