% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_phenotypes_by_date_stages}
\alias{.calc_phenotypes_by_date_stages}
\title{Calculate phenotypes by date stages}
\usage{
.calc_phenotypes_by_date_stages(
  daily_records,
  phenotype_method,
  cfg,
  target_date_stages
)
}
\description{
Calculate phenotypes by date stages
}
\keyword{internal}
