% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_phenotypes_by_weight_stages}
\alias{.calc_phenotypes_by_weight_stages}
\title{Calculate phenotypes by weight stages}
\usage{
.calc_phenotypes_by_weight_stages(
  daily_records,
  phenotype_method,
  cfg,
  target_weight_stages
)
}
\description{
Calculate phenotypes by weight stages
}
\keyword{internal}
