% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_phenotypes_monitor}
\alias{.calc_phenotypes_monitor}
\title{Monitor mode: rolling window method}
\usage{
.calc_phenotypes_monitor(daily_records, cfg, window = 7)
}
\description{
Monitor mode: rolling window method
}
\keyword{internal}
