% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_phenotypes_report}
\alias{.calc_phenotypes_report}
\title{Report mode: stage average method}
\usage{
.calc_phenotypes_report(daily_records, cfg)
}
\description{
Report mode: stage average method
}
\keyword{internal}
