% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_phenotypes_research}
\alias{.calc_phenotypes_research}
\title{Research mode: linear regression method}
\usage{
.calc_phenotypes_research(daily_records, cfg)
}
\description{
Research mode: linear regression method
}
\keyword{internal}
