% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.calc_phenotypes_standard_fcr}
\alias{.calc_phenotypes_standard_fcr}
\title{Standard FCR calculation (30-120kg)}
\usage{
.calc_phenotypes_standard_fcr(daily_records, cfg)
}
\description{
Standard FCR calculation (30-120kg)
}
\keyword{internal}
