% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_impute_national.R
\name{.check_fcr_stages_v2}
\alias{.check_fcr_stages_v2}
\title{Check FCR in weight stages (Table 2)}
\usage{
.check_fcr_stages_v2(data, config)
}
\description{
Check FCR in weight stages (Table 2)
}
\keyword{internal}
