% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_phenotype_calc.R
\name{.dispatch_phenotype_calc}
\alias{.dispatch_phenotype_calc}
\title{Dispatch phenotype calculation to specific method}
\usage{
.dispatch_phenotype_calc(daily_records, phenotype_method, cfg)
}
\description{
Dispatch phenotype calculation to specific method
}
\keyword{internal}
