% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_utils.R
\name{.find_tabular_files_recursive}
\alias{.find_tabular_files_recursive}
\title{Recursively find supported tabular data files}
\usage{
.find_tabular_files_recursive(data_path)
}
\description{
Recursively find supported tabular data files
}
\keyword{internal}
