% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_utils.R
\name{.get_mapped_col}
\alias{.get_mapped_col}
\title{Get mapped column name from format metadata}
\usage{
.get_mapped_col(dt, format_info, field_name)
}
\description{
Get mapped column name from format metadata
}
\keyword{internal}
