% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_utils.R
\name{.get_mapped_unit}
\alias{.get_mapped_unit}
\title{Get mapped unit from format metadata}
\usage{
.get_mapped_unit(format_info, field_name)
}
\description{
Get mapped unit from format metadata
}
\keyword{internal}
