% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_qc_utils.R
\name{.identify_bad_dates}
\alias{.identify_bad_dates}
\title{Identify dates with all weight anomalies}
\usage{
.identify_bad_dates(weight_records, flag_col = "flag_outlier_single")
}
\arguments{
\item{weight_records}{Data table with record_date and flag columns}

\item{flag_col}{Flag column name indicating single-record anomalies}
}
\value{
Vector of dates where all records are anomalous
}
\description{
找出某日所有体重都是异常的日期
}
\keyword{internal}
