% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_impute.R
\name{.impute_weight_national}
\alias{.impute_weight_national}
\title{National standard weight imputation (Kalman filter)}
\usage{
.impute_weight_national(daily_records, cfg)
}
\description{
National standard weight imputation (Kalman filter)
}
\keyword{internal}
