% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_data.R
\name{.map_fire_to_standard}
\alias{.map_fire_to_standard}
\title{Map FIRE columns to standard format}
\usage{
.map_fire_to_standard(dt, format_info, birth_info = NULL)
}
\description{
Map FIRE columns to standard format
}
\keyword{internal}
