% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_data.R
\name{.map_yangxiang_to_standard}
\alias{.map_yangxiang_to_standard}
\title{Map YANGXIANG columns to standard format}
\usage{
.map_yangxiang_to_standard(dt, format_info)
}
\description{
Map YANGXIANG columns to standard format
}
\keyword{internal}
