% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_qc_feed_standard.R
\name{.qc_feed_standard_national}
\alias{.qc_feed_standard_national}
\title{National standard feed QC for standard records}
\usage{
.qc_feed_standard_national(standard_records, cfg, logger = NULL)
}
\description{
National standard feed QC for standard records
}
\keyword{internal}
