% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_qc_weight_standard.R
\name{.qc_weight_standard_national}
\alias{.qc_weight_standard_national}
\title{National standard weight QC for standard records}
\usage{
.qc_weight_standard_national(standard_records, cfg, logger = NULL)
}
\description{
National standard weight QC for standard records
}
\keyword{internal}
