% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_utils.R
\name{.read_shared_data_format_file}
\alias{.read_shared_data_format_file}
\title{Read and parse data format file (JSON ONLY)}
\usage{
.read_shared_data_format_file(format_path)
}
\description{
Read and parse data format file (JSON ONLY)
}
\keyword{internal}
