% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_utils.R
\name{.read_yangxiang_file}
\alias{.read_yangxiang_file}
\title{Read single YANGXIANG xlsx file}
\usage{
.read_yangxiang_file(file, format_info)
}
\description{
Read single YANGXIANG xlsx file
}
\keyword{internal}
