% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_logging.R
\name{ZhenM_close_log}
\alias{ZhenM_close_log}
\title{Close log file}
\usage{
ZhenM_close_log(logger)
}
\arguments{
\item{logger}{Logger object from ZhenM_init_log or ZhenM_create_logger}
}
\description{
Close log file
}
