% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_standard_format.R
\name{ZhenM_create_standard_template}
\alias{ZhenM_create_standard_template}
\title{Create empty standard format data.table}
\usage{
ZhenM_create_standard_template()
}
\value{
An empty data.table with standard format columns
}
\description{
Create empty standard format data.table
}
