% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_read_data.R
\name{ZhenM_data_summary}
\alias{ZhenM_data_summary}
\title{Get data summary statistics}
\usage{
ZhenM_data_summary(standard_data)
}
\arguments{
\item{standard_data}{A data.table in standard format}
}
\value{
A list with summary statistics
}
\description{
Get data summary statistics
}
