% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_config_defaults.R
\name{ZhenM_default_config}
\alias{ZhenM_default_config}
\title{Get default configuration for ZhenMeasure}
\usage{
ZhenM_default_config(qc_method = "national_standard")
}
\arguments{
\item{qc_method}{QC method: "national_standard" (the only supported method since V1.0.0)}
}
\value{
A list of configuration parameters
}
\description{
Returns default configuration parameters for quality control, imputation,
and phenotype calculation using the national standard method.
}
\examples{
# Get national standard config
cfg <- ZhenM_default_config("national_standard")
}
