% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_qc_summary.R
\name{ZhenM_generate_qc_summary}
\alias{ZhenM_generate_qc_summary}
\title{Generate QC summary report}
\usage{
ZhenM_generate_qc_summary(qc_data)
}
\arguments{
\item{qc_data}{Data with QC flags}
}
\value{
Summary data.table
}
\description{
Generate QC summary report
}
