% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_impute.R
\name{ZhenM_impute_feed}
\alias{ZhenM_impute_feed}
\title{Impute feed data}
\usage{
ZhenM_impute_feed(
  daily_records,
  impute_method = "national_standard",
  config = NULL
)
}
\description{
Impute feed data
}
\keyword{internal}
