% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_impute.R
\name{ZhenM_impute_weight}
\alias{ZhenM_impute_weight}
\title{Impute weight data}
\usage{
ZhenM_impute_weight(
  daily_records,
  impute_method = "national_standard",
  config = NULL
)
}
\description{
Impute weight data
}
\keyword{internal}
