% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_logging.R
\name{ZhenM_log}
\alias{ZhenM_log}
\title{Write message to log}
\usage{
ZhenM_log(logger, msg)
}
\arguments{
\item{logger}{Logger object from ZhenM_init_log or ZhenM_create_logger}

\item{msg}{Message to log}
}
\description{
Write message to log
}
