% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_logging.R
\name{ZhenM_log_threshold}
\alias{ZhenM_log_threshold}
\title{Log threshold information}
\usage{
ZhenM_log_threshold(logger, threshold_name, value, unit = "")
}
\arguments{
\item{logger}{Logger object}

\item{threshold_name}{Name of the threshold}

\item{value}{Threshold value}

\item{unit}{Optional unit string}
}
\description{
Log threshold information
}
