% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_daily_aggregate_filtered.R
\name{.apply_feed_lmm_correction}
\alias{.apply_feed_lmm_correction}
\title{LMM Feed Intake Correction Engine}
\usage{
.apply_feed_lmm_correction(daily_dt, raw_dt)
}
\description{
LMM Feed Intake Correction Engine
}
\keyword{internal}
