% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_write_outputs.R
\name{.plot_animal_2x2}
\alias{.plot_animal_2x2}
\title{Plot single animal with 2x2 panel layout}
\usage{
.plot_animal_2x2(raw_sub, daily_sub, animal_id)
}
\description{
Row 1: Raw daily data (original before QC)
  Left: record_date vs daily_feed_g (#f5616f=QC Removed, #f7b13f=Partial Anomaly, black=Normal)
  Right: record_date vs daily_weight_g (#f5616f=QC Removed, #f7b13f=Partial Anomaly, black=Normal)
Row 2: Cleaned daily data (after QC and imputation)
  Left: record_date vs daily_feed_g (green triangle=Imputed, #3685fe circle=Corrected, black circle=Normal)
  Right: record_date vs daily_weight_g (green triangle=Imputed, #3685fe circle=Corrected, black circle=Normal)
}
\keyword{internal}
