% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhenm_write_outputs.R
\name{.plot_panel_tricolor}
\alias{.plot_panel_tricolor}
\title{Plot panel with three-color classification (original data)}
\usage{
.plot_panel_tricolor(x, y, flag_imputed, flag_outlier, main, xlab, ylab)
}
\description{
Color scheme: #f5616f = QC Removed, #f7b13f = Partial Anomaly, black = Normal
}
\keyword{internal}
